refactor testing behavior
