+++ test/account/Account.behavior.js
@@ -4,7 +4,6 @@ const { setBalance } = require('@nomicfoundation/hardhat-network-helpers');
 
 const { impersonate } = require('@openzeppelin/contracts/test/helpers/account');
 const { SIG_VALIDATION_SUCCESS, SIG_VALIDATION_FAILURE } = require('@openzeppelin/contracts/test/helpers/erc4337');
-const { CALL_TYPE_BATCH, encodeMode, encodeBatch } = require('@openzeppelin/contracts/test/helpers/erc7579');
 const {
   shouldSupportInterfaces,
 } = require('@openzeppelin/contracts/test/utils/introspection/SupportsInterface.behavior');
@@ -149,145 +148,7 @@ function shouldBehaveLikeAccountHolder() {
   });
 }
 
-function shouldBehaveLikeAccountERC7821({ deployable = true } = {}) {
-  describe('execute', function () {
-    beforeEach(async function () {
-      // give eth to the account (before deployment)
-      await setBalance(this.mock.target, ethers.parseEther('1'));
-
-      // account is not initially deployed
-      await expect(ethers.provider.getCode(this.mock)).to.eventually.equal('0x');
-
-      this.encodeUserOpCalldata = (...calls) =>
-        this.mock.interface.encodeFunctionData('execute', [
-          encodeMode({ callType: CALL_TYPE_BATCH }),
-          encodeBatch(...calls),
-        ]);
-    });
-
-    it('should revert if the caller is not the canonical entrypoint or the account itself', async function () {
-      await this.mock.deploy();
-
-      await expect(
-        this.mock.connect(this.other).execute(
-          encodeMode({ callType: CALL_TYPE_BATCH }),
-          encodeBatch({
-            target: this.target,
-            data: this.target.interface.encodeFunctionData('mockFunctionExtra'),
-          }),
-        ),
-      )
-        .to.be.revertedWithCustomError(this.mock, 'AccountUnauthorized')
-        .withArgs(this.other);
-    });
-
-    if (deployable) {
-      describe('when not deployed', function () {
-        it('should be created with handleOps and increase nonce', async function () {
-          const operation = await this.mock
-            .createUserOp({
-              callData: this.encodeUserOpCalldata({
-                target: this.target,
-                value: 17,
-                data: this.target.interface.encodeFunctionData('mockFunctionExtra'),
-              }),
-            })
-            .then(op => op.addInitCode())
-            .then(op => this.signUserOp(op));
-
-          // Can't call the account to get its nonce before it's deployed
-          await expect(entrypoint.getNonce(this.mock.target, 0)).to.eventually.equal(0);
-          await expect(entrypoint.handleOps([operation.packed], this.beneficiary))
-            .to.emit(entrypoint, 'AccountDeployed')
-            .withArgs(operation.hash(), this.mock, this.factory, ethers.ZeroAddress)
-            .to.emit(this.target, 'MockFunctionCalledExtra')
-            .withArgs(this.mock, 17);
-          await expect(this.mock.getNonce()).to.eventually.equal(1);
-        });
-
-        it('should revert if the signature is invalid', async function () {
-          const operation = await this.mock
-            .createUserOp({
-              callData: this.encodeUserOpCalldata({
-                target: this.target,
-                value: 17,
-                data: this.target.interface.encodeFunctionData('mockFunctionExtra'),
-              }),
-            })
-            .then(op => op.addInitCode());
-
-          operation.signature = '0x00';
-
-          await expect(entrypoint.handleOps([operation.packed], this.beneficiary)).to.be.reverted;
-        });
-      });
-    }
-
-    describe('when deployed', function () {
-      beforeEach(async function () {
-        await this.mock.deploy();
-      });
-
-      it('should increase nonce and call target', async function () {
-        const operation = await this.mock
-          .createUserOp({
-            callData: this.encodeUserOpCalldata({
-              target: this.target,
-              value: 42,
-              data: this.target.interface.encodeFunctionData('mockFunctionExtra'),
-            }),
-          })
-          .then(op => this.signUserOp(op));
-
-        await expect(this.mock.getNonce()).to.eventually.equal(0);
-        await expect(entrypoint.handleOps([operation.packed], this.beneficiary))
-          .to.emit(this.target, 'MockFunctionCalledExtra')
-          .withArgs(this.mock, 42);
-        await expect(this.mock.getNonce()).to.eventually.equal(1);
-      });
-
-      it('should support sending eth to an EOA', async function () {
-        const operation = await this.mock
-          .createUserOp({ callData: this.encodeUserOpCalldata({ target: this.other, value }) })
-          .then(op => this.signUserOp(op));
-
-        await expect(this.mock.getNonce()).to.eventually.equal(0);
-        await expect(entrypoint.handleOps([operation.packed], this.beneficiary)).to.changeEtherBalance(
-          this.other,
-          value,
-        );
-        await expect(this.mock.getNonce()).to.eventually.equal(1);
-      });
-
-      it('should support batch execution', async function () {
-        const value1 = 43374337n;
-        const value2 = 69420n;
-
-        const operation = await this.mock
-          .createUserOp({
-            callData: this.encodeUserOpCalldata(
-              { target: this.other, value: value1 },
-              {
-                target: this.target,
-                value: value2,
-                data: this.target.interface.encodeFunctionData('mockFunctionExtra'),
-              },
-            ),
-          })
-          .then(op => this.signUserOp(op));
-
-        await expect(this.mock.getNonce()).to.eventually.equal(0);
-        const tx = entrypoint.handleOps([operation.packed], this.beneficiary);
-        await expect(tx).to.changeEtherBalances([this.other, this.target], [value1, value2]);
-        await expect(tx).to.emit(this.target, 'MockFunctionCalledExtra').withArgs(this.mock, value2);
-        await expect(this.mock.getNonce()).to.eventually.equal(1);
-      });
-    });
-  });
-}
-
 module.exports = {
   shouldBehaveLikeAccountCore,
   shouldBehaveLikeAccountHolder,
-  shouldBehaveLikeAccountERC7821,
 };
